mido: android.hardware.graphics.allocator@2.0 is no longer VNDK-SP
The interface lib has been in VNDK-SP because
android.hardware.graphics.mapper@1.0 was using it. However, since the
dependency has gone [1], there is no need keep it in VNDK-SP. The
VNDK-SP set should be kept as small as possible because libs in VNDK-SP
are subject to double-loading.

[1] https://android.googlesource.com/platform/hardware/interfaces/+/79d13ff0f5cd3c4242445baf5b2f01a52f0ed06f

Bug: 69480083
Test: /system/lib64/vndk-sp/android.hardware.graphics.allocator@2.0.so
disappears.
Test: sailfish boots

Change-Id: I66ddfc91e6178def39671bd5f90330b4601cfd89
Merged-In: I66ddfc91e6178def39671bd5f90330b4601cfd89